Add MCP chat client with full conversation pipeline
- ChatMCPClient: interactive chat that routes natural language to MCP
  tools (review, email, summary, list, date, tip, free_chat)
- MCPToolRegistry: registered free_chat, java_tip, get_date_time,
  save/load/list_report, email_report_content tools
- MCPServer: removed 30s socket timeout for long Ollama calls
- Keyword fast-path in intent classifier so single-word commands
  (email, review, tip, etc.) never mis-route through the LLM
- Session transcript written to chat-logs/ with full MCP
  request/response details for later review
- Star-line separators in log file around each user turn for
  easy log reading
- logback: WARN+ appears in chat window; INFO goes to file only
- .gitignore: fixed broken .DS_Store/logs/ merge; added chat-logs/
  and chat-reports/ to exclusions

+++ src/main/java/com/agentic/codereview/rag/EnhancedVectorStore.java
@@ -67,7 +67,7 @@ public class EnhancedVectorStore {
     private float computeBM25(EmbeddingVector query, EmbeddingVector doc) {
         final float k1 = 1.5f;    // Term frequency saturation
         final float b = 0.75f;    // Length normalization
-        final float avgDocLength = vectors.values().stream()
+        final float avgDocLength = (float) vectors.values().stream()
                 .mapToInt(EmbeddingVector::getLength)
                 .average()
                 .orElse(100);
